demo-marroquineria/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
distlib==0.3.6
Django==4.1.7
django-environ==0.10.0
django-postgresql==0.0.3
djangorestframework==3.14.0
env==0.1.0
filelock==3.12.2
Pillow==9.4.0
pipenv==2023.6.12
platformdirs==3.5.3
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.23.0
virtualenv-clone==0.5.7
